Updated file paths and part of the script
Updated file paths but also changed part of the script. Those lme functions were getting a "false convergence" which means there was a problem with the derivative calculation. Since I didn't know enough about the function of the exact goals I fiddled to get it to work and switching the order fixed it. I think it still does what the original was supposed to but I need Christy to check it